background_control 분리
